repoman: Merge USEFlagChecks and UnusedCheck into PkgMetadata class
This eliminates muselist and used_useflags from dynamic_data.
